+++ tests/test_extensions.py
@@ -27,7 +27,9 @@ class DummyClass:
     dummy: Greeter  # available when using `dummy_namespace` fixture
 
 
-register_dummy_namespace = make_register_namespace_decorator(DummyClass, "obj", "register_dummy_namespace")
+register_dummy_namespace = make_register_namespace_decorator(
+    DummyClass, "obj", "register_dummy_namespace", docstring_style="google"
+)
 
 
 @pytest.fixture